+++ RELEASE-NOTES.md
@@ -16,6 +16,8 @@ The project release numbers follow [Semantic Versioning](http://semver.org/spec/
   the `clang-apply-replacements` executable in addition to the `CLANGTIDY_EXECUTABLE`.
   Also adds a corresponding `ENABLE_CLANGAPPLYREPLACEMENTS` CMake option.
   Note that the `clang_tidy_style` target is not added to the `style` target and must be run separately.
+- Added the `blt_install_tpl_setups` macro, which installs files to setup and create targets for the third-party libraries OpenMP, MPI, CUDA, and HIP.  This macro is meant to 
+replace `blt_export_tpl_targets` as the preferred way to setup third-party libraries with BLT.
 
 ### Changed
 - SetupHIP now searches for user-defined or environment variables before CMake paths to find the ROCM_PATH.